Added false check for glob() in listPosts()
On some systems when no .xml files are found in the blog posts folder, glob() will return 'false' instead of an empty array. This is fixed by adding an extra condition to the 'if' statement that follows that will check for 'false' from glob() and cause the listPosts() function to also return 'false' like it is supposed to when there are no .xml post files.