Added better handling of having no posts
On some systems, glob() will return false instead of an empty array when
there are no post .xml files. This is fixed by checking for the false
condition and converting it to an empty array. An if condition has been
added prior to the foreach loop to check if the array is empty, and if
it is, return the feed with a single post saying that the feed is empty.