Fixed semantic issues with post count in categories
Fixed a number of semantic issues with category post counts that were
found during testing of the function. This includes a couple of missing
punctuation characters, casting of XMLobjects to string, an illegal
offset in show_blog_category() [the array this function counts], plus
the check for enabling counts was changed to use is_bool().

Part of #074