Fixed language string concatenation
The strings 'Manage' and 'Custom Fields' for the title of the relevant
admin page were concatenated. 'CUSTOM_FIELDS' is not used anywhere else,
so 'MANAGE' was removed. This means that language file don't need a new
string added to them and can likely use the existing translated string
from 'CUSTOM_FIELDS' for the title of the page. Translators may change
this string if they feel that it is needed.

Fixes #076